% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTMRmain.R
\name{LTMcook}
\alias{LTMcook}
\title{Prepare the data for running LTMdish}
\usage{
LTMcook(cookD, corMethod = "pearson", outFileSymbol = NULL, outDir = NULL)
}
\arguments{
\item{cookD}{a data frame. The screen data from LTMheat.}

\item{corMethod}{a character string. Select the correlation coefficient is to be used for
the correlation test between expression level and measures of clock strength.
It must be \code{"pearson"}(default), \code{"kendall"}, or \code{"spearman"}.}

\item{outFileSymbol}{a character string. A prefix exists in the name of output file.
Set it \code{NULL}(default) if prefer to return a data frame.}

\item{outDir}{a character string. The name of directory used to store output file.}
}
\value{
a data fame
}
\description{
This function calculates the correlation value between expression level and measures of clock strength, gene by gene.
The input screen data is from LTMheat.
}
\examples{
## please refer to the webpage of LTMR package
}
