% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTMRmain.R
\name{LTMcut}
\alias{LTMcut}
\title{Prepare the input data for a quick run of LTMheat}
\usage{
LTMcut(
  cutD = NULL,
  exemptGenes = NULL,
  minExp = -1e+10,
  minFold = 1,
  outFileSymbol = NULL,
  outDir = NULL
)
}
\arguments{
\item{cutD}{a data frame. The first column is the gene symbol, and other columns are samples. One row per gene.}

\item{exemptGenes}{a character vector. A list of clock genes that will be kept in the output data frame without getting into the filtration process.}

\item{minExp}{a numeric value. The minimal median expression value in quantile group 1.}

\item{minFold}{a numeric value. The minimal fold change between quantile group 4 and quantile group 1. Median value in each quantile group is used.}

\item{outFileSymbol}{a character string. A prefix exists in the name of output file.
Set it \code{NULL}(default) if prefer to return a data frame.}

\item{outDir}{a character string. The name of directory used to store output file.}
}
\value{
a data fame prepared for a quick run of LTM
}
\description{
This function will select genes with high expression values and large expression variations across samples for a quick run of LTMheat.
}
\examples{
## please refer to the webpage of LTMR package
}
