% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTMRmain.R
\name{LTMdish}
\alias{LTMdish}
\title{Generate the LTM analysis result for the input expression dataset}
\usage{
LTMdish(
  dishL,
  targetMeasures = c("zmantel", "zncv"),
  onlyConsistentTrend = FALSE,
  fileName = NULL,
  outDir = NULL
)
}
\arguments{
\item{dishL}{a list of data frame. Each data frame is from LTMcook output.}

\item{targetMeasures}{a vector of character string. The target measures of clock strength used to calculate LTM_abs value for each gene.}

\item{onlyConsistentTrend}{logical. Set \code{TRUE} will only keep those genes show consistent trend across targetMeasures.}

\item{fileName}{a character string. The name of output file. Set it \code{NULL}(default) if prefer to return a data frame.}

\item{outDir}{a character string. The name of directory used to store output file.}
}
\value{
a data fame
}
\description{
This function calculates the LTM_abs value for each gene, which indicates the correlation between gene expression and clock strength.
}
\examples{
## please refer to the webpage of LTMR package
}
