% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTMRmain.R
\name{LTMheat}
\alias{LTMheat}
\title{Prepare the screen and sample data used for running LTMcook}
\usage{
LTMheat(
  heatD,
  benchD,
  cvGenes,
  outFileSymbol = NULL,
  outDir = NULL,
  qnum = 4,
  nCores = 1,
  releaseNote = TRUE
)
}
\arguments{
\item{heatD}{a data frame. The first column is the gene symbol, and other columns are samples. One row per gene.}

\item{benchD}{a data frame. The expression correlation values of paired clock genes.}

\item{cvGenes}{a vector string. The list of clock genes representing the oscillation robustness of circadian clock.}

\item{outFileSymbol}{a character string. A prefix exists in the name of output file.
Set it \code{NULL}(default) if prefer to return a data frame.}

\item{outDir}{a character string. The name of directory used to store output file.}

\item{qnum}{a numeric value. The number of quantile groups when separate samples based on expression level gene by gene.}

\item{nCores}{an integer. Number of cores to use. For parallel computing, set \code{nCores} a integer larger than 1.}

\item{logical.}{If TRUE, time used during the analysis will be released on the screen.}
}
\value{
a list with screen and sample data frame.
}
\description{
This function will separate the samples into given number of quantile groups, calculate the mantel's zstat value,
nCV values of given clock gene list, gene by gene. It will output a list with screen and sample data frame used for running LTMcook.
}
\examples{
## please refer to the webpage of LTMR package
}
