% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTMRmain.R
\name{LTMprep}
\alias{LTMprep}
\title{Prepare the input data for LTM analysis}
\usage{
LTMprep(
  prepD,
  quantNorm = TRUE,
  uniStyle = "mad",
  removeLowQuant = 0.1,
  bluntLowQuant = 0.025,
  bluntHighQuant = 0.975,
  digitsNum = 4,
  outFileSymbol = NULL,
  outDir = NULL
)
}
\arguments{
\item{prepD}{a data frame. The first column is the gene symbol, and other columns are samples. One row per gene.}

\item{quantNorm}{logical. Set \code{TRUE} will do quantile normalization across samples.}

\item{uniStyle}{a character string. Select the way of getting the representative profile for rows with duplicated gene symbols.
It must be \code{"mad"}(default), \code{"sum"}, or \code{"none"}, represent 'select the row with max mean absolute deviation', 'sum up duplicate rows'
or 'there is no duplicate gene name, and skip the step of getting the representative profile'.}

\item{removeLowQuant}{a numeric value. Filter out genes based on their median expression value. If the median expression value locate below the \code{0.25}(default) quantile,
this gene will be filtered out. Set it to \code{0} for keeping all the input genes.}

\item{bluntLowQuant}{a numeric value. Blunt the outlier value across samples gene by gene. If an expression value is below the \code{0.025}(default) quantile,
this outlier value will be blunted.}

\item{bluntHighQuant}{a numeric value. Blunt the outlier value across samples gene by gene. If an expression value is above the \code{0.975}(default) quantile,
this outlier value will be blunted.}

\item{digitsNum}{an integer value. The integer indicates the number of decimal places to be used in the prepared data frame.}

\item{outFileSymbol}{a character string. A prefix exists in the name of output file.
Set it \code{NULL}(default) if prefer to return a data frame.}

\item{outDir}{a character string. The name of directory used to store output file.}
}
\value{
a data fame prepared for LTM analysis
}
\description{
This function can do quantile normalization, get the representative profile for rows with duplicated gene symbols,
filter out genes with low expression values, and blunt the outliers of the given expression matrix.
}
\examples{
## please refer to the webpage of LTMR package
}
