% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTMRmain.R
\name{LTMwash}
\alias{LTMwash}
\title{Delete unnecessary files}
\usage{
LTMwash(washDir, washKey = NULL, checkWashList = FALSE)
}
\arguments{
\item{washDir}{a character string. The name of directory used to store intermediate files during LTM analysis.}

\item{washKey}{a character string. The shared key word among targeted files under deletion.}

\item{checkWashList}{logical. Set \code{TRUE} will return a list of files.}
}
\value{
a data frame of deleted files.
}
\description{
This function helps delete intermediate files generated during LTM analysis.
}
\examples{
## please refer to the webpage of LTMR package
}
